feat: refactor, merge running and compelted call graphs, render merged flamegraphs in svg
