+++ crates/telemetry-subscribers/src/flamegraph/flame.rs
@@ -242,3 +242,40 @@ impl<S: Default + MergeMetrics + SpanMetrics> Flames<S> {
         }
     }
 }
+impl<S: Clone + Default + MergeMetrics + SpanMetrics> Flames<S> {
+    /// Merge and collect a running call graph into completed with the given
+    /// graph id.
+    pub fn get_callgraph(&self, graph_id: &GraphId) -> Option<CallGraph<S>> {
+        let running = self
+            .graphs
+            .read()
+            .values()
+            .find(|g| g.graph_id == *graph_id)
+            .map(|g| g.mutex.lock().clone());
+        let mut completed = self.completed.read().get(graph_id).cloned();
+        if let Some(g) = &mut completed {
+            if let Some(running) = running {
+                g.merge(running);
+            }
+        }
+        completed
+    }
+    /// Merge and collect all running call graphs into completed with the same
+    /// graph id.
+    pub fn get_callgraphs(&self) -> HashMap<GraphId, CallGraph<S>> {
+        let mut callgraphs = self.completed.read().clone();
+        let rlock = self.graphs.read();
+        rlock.iter().for_each(|(_tid, Graph { graph_id, mutex })| {
+            // collect_into is still unstable
+            match callgraphs.entry(*graph_id) {
+                hash_map::Entry::Occupied(mut entry) => {
+                    entry.get_mut().merge(mutex.lock().clone());
+                }
+                hash_map::Entry::Vacant(entry) => {
+                    entry.insert(mutex.lock().clone());
+                }
+            }
+        });
+        callgraphs
+    }
+}
